Reworked I am looking for RPA page
With the announcement by Robocorp and the uncertainity around using
their full suite of tools we want to not misdirect users on a solution
that might not be be there in the near future. As such I have updated
the I am looking for RPA page. It removes most of the Robocorp content.
It does leave reference to RPAFramework libraries which as a temporary
change I am leaving.