Update style_guide.md
In the Lists section, there is an example to demonstrate how to use the Long item in the list. one long item variable is added as the last element but the syntax should be ${LONG ITEM} but it is $(LONG ITEM) which doesn't expand to the value that LONG ITEM variable stores.